viz/requirements.txt
streamlit>=1.28.0
streamlit-folium>=0.15.0
folium>=0.15.0
geopandas>=0.14.0
sqlalchemy>=2.0
psycopg2-binary
geoalchemy2>=0.14.0
branca>=0.7.0
requests>=2.31.0
rasterio>=1.3.0
Pillow>=10.0.0
